+++ StreamChatUITestsAppUITests/Pages/ChannelListPage.swift
@@ -45,9 +45,9 @@ enum ChannelListPage {
         }
 
         static func statusCheckmark(for status: MessageDeliveryStatus?, in cell: XCUIElement) -> XCUIElement {
-            var identifier = "imageView"
+            var identifier = "There is no status checkmark"
             if let status = status {
-                identifier = "\(identifier)_\(status.rawValue)"
+                identifier = "imageView_\(status.rawValue)"
             }
             return cell.images[identifier]
         }